Address code review of the small_changes branch
Review fixes:

collide_vss_kokkos: collisions_one_subcell() was never updated for the
  d_scalars/d_scalars_big split, so it only deep-copied the int scalars.
  With atomic_reduction on (GPU), the subcell path reported Natt, Ncoll,
  Nreact and the end-of-run collide totals as zero.  Verified by forcing
  atomic_reduction on for a host build: before, in.collide_3D_sub
  reports 0 attempts and 0 collisions; after, it reports the same counts
  as the parallel_reduce path.
collide_vss_kokkos: give collisions_one_subcell() the same react_extra
  overflow guards the other two collision routines received.
fix_ablate: stop the decrement/sync retry loop as soon as a pass pays
  nothing more, instead of always running all MAXDECITER passes when a
  cell owes a decrement no corner point can cover.  Every pass past the
  first unproductive one is a no-op that still costs a
  comm_neigh_corners() plus an MPI_Allreduce.  in.ablation.3d Modify
  time 64.7 s -> 55.7 s, with the log unchanged.
particle_kokkos: k_map.resize(extent*1.5) truncates back to the old size
  when the view holds one element, so h_map[nlocal] wrote one past the
  end for a proc owning a single particle with ratio >= 2.
irregular: uniform_datum_type() returned an uninitialized datum_type
  when called with nbytes = 0, since 0 was also the "none built yet"
  marker.  Use -1 as the marker instead.

Review feedback:

Drop the LAMMPS reference from the sparta_extract_setting() docs.
Show MPI Sync in the screen-output example in Section_start and remove
  the note that the example predates it.
Drop the silent MIN(...,MAXSMALLINT) clamps whose quantity cannot
  realistically reach 2^31: maxcellcount_extra and maxelectron_extra in
  the Kokkos collisions, the 5 per-surf and per-face adsorbate counts in
  surf_react_adsorb, and nsplit/nsub in the restart header, which count
  only the grid cells a surface cuts and whose clamped value would
  silently defeat the check that reads them back.  The clamp at zero in
  surf_react_adsorb stays: net desorption really can drive a count
  negative, which is what the pre-existing MAX(0,...) is for.
Where the quantity can reach 2^31, error out instead of clamping:
  maxdelete_extra in the Kokkos collisions, and maxdot in rcb, where a
  clamped growth request leaves the dot array too small and the code
  then writes past the end of it.
fix_grid_check_kokkos: report the count of particles in wrong cells as a
  bigint, matching the non-Kokkos fix, instead of clamping it to fit %d.
Use bigint rather than long, long long or int64_t for the 64-bit
  intermediates and for crs_size_type.
Revert the MPI stubs to int byte counts: only one proc exists there, so
  a message that overflows an int cannot be allocated in the first
  place.
Revert nstuck and naxibad to int: both count rare events.  The Kokkos
  side already keeps them in the int scalar view and in UPDATE_REDUCE as
  ints, so this makes the two agree again.

+++ src/update.h
@@ -75,8 +75,8 @@ class Update : protected Pointers {
   bigint nscheck_running;
   bigint nscollide_running;
 
-  bigint nstuck;             // # of particles stuck on surfs and deleted
-  bigint naxibad;            // # of particles where axisymm move was bad
+  int nstuck;                // # of particles stuck on surfs and deleted
+  int naxibad;               // # of particles where axisymm move was bad
                              // in this case, bad means particle ended up
                              // outside of final cell curved surf by epsilon
                              // when move logic thinks it is inside cell
